+++ gsa/src/web/pages/portlists/row.js
@@ -38,8 +38,6 @@ import IconDivider from 'web/components/layout/icondivider';
 import TableData from 'web/components/table/data';
 import TableRow from 'web/components/table/row';
 
-import {PORT_LISTS_FROM_FEED} from 'web/pages/extras/feedstatuspage';
-
 const IconActions = ({
   entity,
   onPortListDeleteClick,
@@ -56,7 +54,7 @@ const IconActions = ({
     />
     <EditIcon
       entity={entity}
-      disabled={PORT_LISTS_FROM_FEED.includes(entity.id)}
+      disabled={entity.predefined}
       displayName={_('Port List')}
       name="port_list"
       onClick={onPortListEditClick}