+++ gsa/src/web/pages/extras/feedstatuspage.js
@@ -78,28 +78,19 @@ const ToolBarIcons = () => (
 const renderCheck = feed => {
   const age = feed.age.asDays();
 
-  return age >= 10
+  return age >= 30 && !hasValue(feed.currentlySyncing)
     ? _('Please check the automatic synchronization of your system.')
     : '';
 };
 
 const renderFeedStatus = feed => {
   if (hasValue(feed.currentlySyncing)) {
-    return (
-      <span style={{paddingLeft: '10px'}}>
-        <img
-          data-testid="loading-indicator"
-          src="/img/loading.gif"
-          title={_('Update in progress')}
-          alt={_('Loading Indicator')}
-        ></img>
-      </span>
-    );
+    return _('Update in progress...');
   }
 
   const age = parseInt(feed.age.asDays());
 
-  if (age >= 10) {
+  if (age >= 30) {
     return _('Too old ({{age}} days)', {age});
   }
 
@@ -107,42 +98,6 @@ const renderFeedStatus = feed => {
     return _('{{age}} days old', {age});
   }
   return _('Current');
-};
-
-export const COMPLIANCE_POLICIES_FROM_FEED = [
-  'c4b7c0cb-6502-4809-b034-8e635311b3e6', // IT-Grundschutz
-];
-
-export const PORT_LISTS_FROM_FEED = [
-  '33d0cd82-57c6-11e1-8ed1-406186ea4fc5', // All IANA assigned TCP
-  '4a4717fe-57d2-11e1-9a26-406186ea4fc5', // All IANA assigned TCP and UDP
-  '730ef368-57e2-11e1-a90f-406186ea4fc5', // All TCP and Nmap top 100 UDP
-];
-
-export const REPORT_FORMATS_FROM_FEED = [
-  '5057e5cc-b825-11e4-9d0e-28d24461215b', // Anonymous XML
-  'c1645568-627a-11e3-a660-406186ea4fc5', // CSV Results
-  '77bd6c4a-1f62-11e1-abf0-406186ea4fc5', // ITG
-  'c402cc3e-b531-11e1-9163-406186ea4fc5', // PDF
-  'a3810a62-1f62-11e1-9219-406186ea4fc5', // TXT
-  'a994b278-1f62-11e1-96ac-406186ea4fc5', // XML
-];
-
-export const SCAN_CONFIGS_FROM_FEED = [
-  'd21f6c81-2b88-4ac1-b7b4-a2a9f2ad4663', // Base
-  '8715c877-47a0-438d-98a3-27c7a6ab2196', // Discovery
-  '085569ce-73ed-11df-83c3-002264764cea', // empty
-  'daba56c8-73ec-11df-a475-002264764cea', // Full and fast
-  '2d3f051c-55ba-11e3-bf43-406186ea4fc5', // Host Discovery
-  'bbca7412-a950-11e3-9109-406186ea4fc5', // System Discovery
-];
-
-export const composeObjFilter = (objectIds = []) => {
-  let filterString = '';
-
-  objectIds.forEach(id => (filterString += 'uuid=' + id + ' '));
-
-  return filterString;
 };
 
 const FeedStatus = ({feeds}) => {
@@ -216,37 +171,25 @@ const FeedStatus = ({feeds}) => {
                   )}
                   {feed.feed_type === GVMD_DATA_FEED && (
                     <IconDivider>
-                      <Link
-                        to="policies"
-                        filter={composeObjFilter(COMPLIANCE_POLICIES_FROM_FEED)}
-                      >
+                      <Link to="policies" filter="predefined=1">
                         <IconDivider align={['start', 'center']}>
                           <PolicyIcon size="medium" />
                           <span>{_('Compliance Policies')}</span>
                         </IconDivider>
                       </Link>
-                      <Link
-                        to="portlists"
-                        filter={composeObjFilter(PORT_LISTS_FROM_FEED)}
-                      >
+                      <Link to="portlists" filter="predefined=1">
                         <IconDivider align={['start', 'center']}>
                           <PortListIcon size="medium" />
                           <span>{_('Port Lists')}</span>
                         </IconDivider>
                       </Link>
-                      <Link
-                        to="reportformats"
-                        filter={composeObjFilter(REPORT_FORMATS_FROM_FEED)}
-                      >
+                      <Link to="reportformats" filter="predefined=1">
                         <IconDivider align={['start', 'center']}>
                           <ReportFormatIcon size="medium" />
                           <span>{_('Report Formats')}</span>
                         </IconDivider>
                       </Link>
-                      <Link
-                        to="scanconfigs"
-                        filter={composeObjFilter(SCAN_CONFIGS_FROM_FEED)}
-                      >
+                      <Link to="scanconfigs" filter="predefined=1">
                         <IconDivider align={['start', 'center']}>
                           <ScanConfigIcon size="medium" />
                           <span>{_('Scan Configs')}</span>
@@ -260,7 +203,7 @@ const FeedStatus = ({feeds}) => {
                 <TableData>
                   <Divider wrap>
                     <strong>{renderFeedStatus(feed)}</strong>
-                    <span>{renderCheck(feed)}</span>
+                    <span data-testid="update-msg">{renderCheck(feed)}</span>
                   </Divider>
                 </TableData>
               </TableRow>