+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 ### Removed
+- Business Process Map (BPM) feature [#2382](https://github.com/greenbone/gsa/pull/2382)
 
 [20.8.1]: https://github.com/greenbone/gsa/compare/v20.8.0...gsa-20.08