+++ CHANGELOG.md
@@ -14,8 +14,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 ### Removed
+- Removed Internet Explorer 11 support
 - Removed parsing and testing of textExcerpt [#2316](https://github.com/greenbone/gsa/pull/2316)
-- Remove extra parsing of comment and summary in Model [#2309](https://github.com/greenbone/gsa/pull/2309)
+- Removed extra parsing of comment and summary in Model [#2309](https://github.com/greenbone/gsa/pull/2309)
 
 [21.04]: https://github.com/greenbone/gsa/compare/gsa-20.08...master